added feature `DatabaseTrait` and moved trait into `databases.rs` file and run cargo fmt
