+++ src/databases/mysql.rs
@@ -1,4 +1,5 @@
-use crate::{AxumDatabasePool, SessionError};
+use super::databases::AxumDatabasePool;
+use crate::SessionError;
 use async_trait::async_trait;
 use chrono::Utc;
 use sqlx::{pool::Pool, MySql, MySqlPool};
