+++ src/session_store.rs
@@ -1,4 +1,4 @@
-use crate::databases::database::AxumDatabasePool;
+use crate::databases::databases::AxumDatabasePool;
 use crate::{AxumSession, AxumSessionConfig, AxumSessionData, AxumSessionTimers, SessionError};
 use chrono::{Duration, Utc};
 use core::fmt;
